scripts/requirements.txt
commonroad-io>=2024.1
numpy
requests
